requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0
fastapi
uvicorn
pydantic
pymongo
httpx
